Finished calculator functionality
The way it works is through an event listener on the calculator
div. It registers the className of the subdiv and operates according
to the div pressed. For opperators, they depend on wether there already
are 2 nums defined or not.

Some keys, like the golden three and the equal key, have special
behaviors to avoid bugs.